requirements.txt
Pillow
opencv-python
numpy
screeninfo
string
datetime
time
winsound
scikit-learn
pickle
dlib